[client] extract community_spec module
